refactor: isolate go-spatial types
